backend/requirements.txt
# Grandby Backend - Python Dependencies
# Python 3.12

# ==================== FastAPI Framework ====================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# ==================== Database ====================
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# ==================== Authentication & Security ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1

# ==================== HTTP Clients ====================
httpx==0.27.2
aiohttp==3.10.5

# ==================== AI/ML APIs ====================
openai==1.51.0
tiktoken==0.7.0

# ==================== Twilio (Voice Calls) ====================
twilio==9.3.0

# ==================== AWS Services ====================
boto3==1.35.29
botocore==1.35.29

# ==================== Task Queue ====================
celery==5.4.0
redis==5.1.1

# ==================== Monitoring & Logging ====================
sentry-sdk[fastapi]==2.15.0

# ==================== Testing ====================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# ==================== Development Tools ====================
black==24.8.0
flake8==7.1.1
mypy==1.11.2
ipython==8.27.0

# ==================== Data Validation ====================
email-validator==2.2.0

# ==================== Date/Time ====================
python-dateutil==2.9.0
pytz==2024.2

